backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "curaweave-backend"
version = "0.1.0"
description = "FastAPI backend for CuraWeave"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "gunicorn>=21.2.0",
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "pydantic[email]>=2.7.4",
    "pydantic-settings>=2.2.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "python-multipart>=0.0.9",
    "celery>=5.4.0",
    "redis>=5.0.4",
    "stripe>=9.6.0",
    "twilio>=9.1.1",
    "google-genai>=1.0.0",
    "langgraph>=0.1.5",
    "langchain-google-genai>=1.0.5",
    "langchain>=0.2.5",
    "langchain-core>=0.2.9",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "apscheduler>=3.10.4"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
